clean things up
