+++ src/metrics.rs
@@ -1,74 +1,21 @@
-use std::{net::SocketAddr, sync::Arc, time::Duration};
+use std::net::SocketAddr;
 
 use eyre::Result;
-use metrics::{Counter, Histogram, counter, histogram};
-use metrics_derive::Metrics;
 use metrics_exporter_prometheus::PrometheusBuilder;
 use metrics_util::layers::{PrefixLayer, Stack};
+use tokio::net::TcpListener;
+use tracing::{error, info};
 
-use crate::{Args, init_metrics_server};
+use crate::Args;
 
-#[derive(Metrics)]
-#[metrics(scope = "rpc")]
-pub struct ServerMetrics {
-    // Builder proxy metrics
-    #[metric(describe = "Latency for builder client forwarded rpc calls (excluding the engine api)", labels = ["method"])]
-    #[allow(dead_code)]
-    pub builder_forwarded_call: Histogram,
+use http::StatusCode;
+use hyper::service::service_fn;
+use hyper::{Request, Response, server::conn::http1};
+use hyper_util::rt::TokioIo;
+use jsonrpsee::http_client::HttpBody;
+use metrics_exporter_prometheus::PrometheusHandle;
 
-    #[metric(describe = "Number of builder client rpc responses", labels = ["http_status_code", "rpc_status_code", "method"])]
-    #[allow(dead_code)]
-    pub builder_rpc_response_count: Counter,
-
-    // L2 proxy metrics
-    #[metric(describe = "Latency for l2 client forwarded rpc calls (excluding the engine api)", labels = ["method"])]
-    #[allow(dead_code)]
-    pub l2_forwarded_call: Histogram,
-
-    #[metric(describe = "Number of l2 client rpc responses", labels = ["http_status_code", "rpc_status_code", "method"])]
-    #[allow(dead_code)]
-    pub l2_rpc_response_count: Counter,
-}
-
-impl ServerMetrics {
-    pub fn record_builder_forwarded_call(&self, latency: Duration, method: String) {
-        histogram!("rpc.builder_forwarded_call", "method" => method).record(latency.as_secs_f64());
-    }
-
-    pub fn increment_builder_rpc_response_count(
-        &self,
-        http_status_code: String,
-        rpc_status_code: Option<String>,
-        method: String,
-    ) {
-        counter!("rpc.builder_response_count",
-            "http_status_code" => http_status_code,
-            "rpc_status_code" => rpc_status_code.unwrap_or("".to_string()),
-            "method" => method,
-        )
-        .increment(1);
-    }
-
-    pub fn record_l2_forwarded_call(&self, latency: Duration, method: String) {
-        histogram!("rpc.l2_forwarded_call", "method" => method).record(latency.as_secs_f64());
-    }
-
-    pub fn increment_l2_rpc_response_count(
-        &self,
-        http_status_code: String,
-        rpc_status_code: Option<String>,
-        method: String,
-    ) {
-        let mut labels = vec![("http_status_code", http_status_code), ("method", method)];
-        if let Some(rpc_status_code) = rpc_status_code {
-            labels.push(("rpc_status_code", rpc_status_code));
-        }
-
-        counter!("rpc.l2_response_count", &labels).increment(1);
-    }
-}
-
-pub(crate) fn init_metrics(args: &Args) -> Result<Option<Arc<ServerMetrics>>> {
+pub(crate) fn init_metrics(args: &Args) -> Result<()> {
     if args.metrics {
         let recorder = PrometheusBuilder::new().build_recorder();
         let handle = recorder.handle();
@@ -81,8 +28,43 @@ pub(crate) fn init_metrics(args: &Args) -> Result<Option<Arc<ServerMetrics>>> {
         let metrics_addr = format!("{}:{}", args.metrics_host, args.metrics_port);
         let addr: SocketAddr = metrics_addr.parse()?;
         tokio::spawn(init_metrics_server(addr, handle)); // Run the metrics server in a separate task
-        Ok(Some(Arc::new(ServerMetrics::default())))
-    } else {
-        Ok(None)
+    }
+    Ok(())
+}
+
+async fn init_metrics_server(addr: SocketAddr, handle: PrometheusHandle) -> eyre::Result<()> {
+    let listener = TcpListener::bind(addr).await?;
+    info!("Metrics server running on {}", addr);
+
+    loop {
+        match listener.accept().await {
+            Ok((stream, _)) => {
+                let handle = handle.clone(); // Clone the handle for each connection
+                tokio::task::spawn(async move {
+                    let service = service_fn(move |_req: Request<hyper::body::Incoming>| {
+                        let response = match _req.uri().path() {
+                            "/metrics" => Response::builder()
+                                .header("content-type", "text/plain")
+                                .body(HttpBody::from(handle.render()))
+                                .unwrap(),
+                            _ => Response::builder()
+                                .status(StatusCode::NOT_FOUND)
+                                .body(HttpBody::empty())
+                                .unwrap(),
+                        };
+                        async { Ok::<_, hyper::Error>(response) }
+                    });
+
+                    let io = TokioIo::new(stream);
+
+                    if let Err(err) = http1::Builder::new().serve_connection(io, service).await {
+                        error!(message = "Error serving metrics connection", error = %err);
+                    }
+                });
+            }
+            Err(e) => {
+                error!(message = "Error accepting connection", error = %e);
+            }
+        }
     }
 }
